refactor: separate in files & start with unit status
