+++ integrations/sync/client/flows/unit-status.ts
@@ -1,0 +1,10 @@
+/**
+ * unit status flow
+ */
+
+emit("chat:addSuggestion", "/sn-active-unit", "This will show your active unit's name and status.");
+emit(
+  "chat:addSuggestion",
+  "/sn-set-status",
+  "This will open a menu and will allow you to select a status for your active unit.",
+);
